+++ crates/tycho-execution/contracts/scripts/deploy-executors.js
@@ -35,9 +35,10 @@ const deploy_protocols = {
         "rfq:metric",
         "pricelevelstream",
         "rfq:native",
-        "propammfallback",
         "sky",
         "lido_v4",
+        "etherfi",
+        "fallback",
     ],
     "base": [
         "uniswap_v2",
@@ -49,7 +50,8 @@ const deploy_protocols = {
         "native_wrapper",
         "lunarbase",
         "rfq:metric",
-        "rfq:native"
+        "rfq:native",
+        "fallback",
     ],
     "unichain": [
         "uniswap_v2",